ai-service/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
xgboost>=2.0.3
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
redis>=5.0.1
sqlalchemy>=2.0.25
alembic>=1.13.1
googlemaps>=4.10.0
openai>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
httpx>=0.27.0
pillow>=10.2.0
opencv-python>=4.9.0.80
imagehash>=4.3.1
geopy>=2.4.1
shapely>=2.0.2
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black>=24.2.0
ruff>=0.3.0
mypy>=1.8.0
